fix(medium): remediate 46 medium-severity audit findings [F-053..F-098]
Wave 3 of the v0.0.65 full-audit remediation (46 findings, 41 file-groups; 36
fully fixed in-file, the rest have cross-file follow-ups noted below). Backend:
full module builds, vets, and all changed-package tests pass (handler, auth,
service, repository, middleware, migrations, recharge, user, cmd/api, webhook,
ai, onboarding, team, workspace). Frontend: tsc + production build clean.

Highlights:
- F-053 chunked-upload 5xx responses no longer leak B2 bucket / object path /
  Postgres constraint text (generic opaque error + server-side log).
- F-054 gallery analytics limit/days params clamped at the handler boundary.
- F-055 contact CSV import body size-capped (413 over limit).
- F-064 TUS PATCH now enforces Upload-Offset (400 absent/malformed, 409 mismatch).
- F-089 gallery token read snapshot via useRef (no per-render localStorage writes).
- F-092 smart-album favorite + face-cluster loading batched via new
  AssetRepo.GetByIDs (N+1 -> single query).
- F-093 AI culling view serves WebP derivatives (was legacy JPEG).
- F-094/F-095/F-096 AI/CRM panels: prop-change data-race guards on fetch effects.
- F-097 messages SSE token query-param now encodeURIComponent-encoded.
- F-067 subscription upgrade failure path hardened (markUpgradeOrderFailed).
- Plus error-masking, validation-bound, and silent-failure fixes across handlers.

Integration repairs (concurrent-agent fallout, fixed under the consolidated gate):
added AssetRepo.GetByIDs, defined markUpgradeOrderFailed, de-duplicated a clashing
captureLog test helper.

Partial / cross-file follow-ups (tracked for a later pass, not regressions):
- F-056/F-070 password max-length + email masking also needed in auth.go /
  admin_user_service.go / dealer_service.go / auth_password_reset_handler.go.
- F-091 album membership N+1 needs a batched backend endpoint + api-client fn.
- F-098 full CSP nonce needs middleware.ts + layout.tsx wiring (config-side ready).

Frontend vitest: pre-existing main baseline has 58 failing tests; these are
unrelated to this wave (production builds clean) and tracked separately.

+++ backend/internal/database/migrations/027_create_m6_coupon_tables.up.sql
@@ -46,7 +46,12 @@ CREATE TABLE IF NOT EXISTS coupon_redemptions (
     workspace_id UUID NOT NULL REFERENCES workspaces(id),
     state_id INTEGER REFERENCES states(id),
     invoice_id UUID,
-    discount_applied DECIMAL(10,2) NOT NULL,
+    -- Monetary amount in integer paisa (BIGINT), matching the BIGINT-paisa
+    -- billing convention used throughout (migration 022: subtotal_paisa,
+    -- total_paisa, amount_paisa). The Go redemption path already inserts an
+    -- int64 paisa value here (coupon_validation_service.go). Migration 124
+    -- converts already-applied DBs from the legacy DECIMAL(10,2) rupee form. (F-065)
+    discount_applied BIGINT NOT NULL,
     redeemed_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
